Refactoring: moved GitHub connection establishment into separate service
