+++ src/main/java/com/lpvs/service/LPVSDetectService.java
@@ -33,26 +33,12 @@ import java.util.Optional;
 @Service
 @Slf4j
 public class LPVSDetectService {
-    @Value("${github.token}")
-    private String GITHUB_AUTH_TOKEN;
-
-    @Value("${github.api.url}")
-    private String GITHUB_API_URL;
-
-    @Value("${github.login}")
-    private String GITHUB_LOGIN;
 
     private String scannerType;
 
     private LPVSScanossDetectService scanossDetectService;
 
-    private static final String GITHUB_LOGIN_PROP_NAME = "github.login";
-    private static final String GITHUB_AUTH_TOKEN_PROP_NAME = "github.token";
-    private static final String GITHUB_API_URL_PROP_NAME = "github.api.url";
-
-    private static final String GITHUB_LOGIN_ENV_VAR_NAME = "LPVS_GITHUB_LOGIN";
-    private static final String GITHUB_AUTH_TOKEN_ENV_VAR_NAME = "LPVS_GITHUB_TOKEN";
-    private static final String GITHUB_API_URL_ENV_VAR_NAME = "LPVS_GITHUB_API_URL";
+    private LPVSGitHubConnectionService gitHubConnectionService;
 
     @Value("${github.pull.request}")
     private String trigger;
@@ -62,32 +48,20 @@ public class LPVSDetectService {
     @Autowired
     public LPVSDetectService(
             @Value("${scanner:scanoss}") String scannerType,
+            LPVSGitHubConnectionService gitHubConnectionService,
             LPVSScanossDetectService scanossDetectService) {
         this.scannerType = scannerType;
+        this.gitHubConnectionService = gitHubConnectionService;
         this.scanossDetectService = scanossDetectService;
     }
 
     @PostConstruct
     private void init() {
-        this.GITHUB_LOGIN =
-                Optional.ofNullable(GITHUB_LOGIN)
-                        .filter(s -> !s.isEmpty())
-                        .orElse(
-                                Optional.ofNullable(System.getenv(GITHUB_LOGIN_ENV_VAR_NAME))
-                                        .orElse(""));
-        this.GITHUB_AUTH_TOKEN =
-                Optional.ofNullable(GITHUB_AUTH_TOKEN)
-                        .filter(s -> !s.isEmpty())
-                        .orElse(
-                                Optional.ofNullable(System.getenv(GITHUB_AUTH_TOKEN_ENV_VAR_NAME))
-                                        .orElse(""));
-        this.GITHUB_API_URL =
-                Optional.ofNullable(GITHUB_API_URL)
-                        .filter(s -> !s.isEmpty())
-                        .orElse(
-                                Optional.ofNullable(System.getenv(GITHUB_API_URL_ENV_VAR_NAME))
-                                        .orElse(""));
         log.info("License detection scanner: " + scannerType);
+    }
+
+    @EventListener(ApplicationReadyEvent.class)
+    public void runOneScan() {
         if (trigger != null && !trigger.equals("")) {
             try {
                 LPVSQueue webhookConfig = this.getInternalQueueByPullRequest(trigger);
@@ -95,12 +69,6 @@ public class LPVSDetectService {
             } catch (Exception e) {
                 e.printStackTrace();
             }
-        }
-    }
-
-    @EventListener(ApplicationReadyEvent.class)
-    public void exitTheApp() {
-        if (trigger != null && !trigger.equals("")) {
             doThisStuffToExit();
         }
     }
@@ -109,13 +77,6 @@ public class LPVSDetectService {
     public void doThisStuffToExit() {
         System.out.println("Exiting...");
         System.exit(0);
-    }
-
-    private static GitHub gitHub;
-
-    public void setGithubTokenFromEnv() {
-        if (System.getenv("LPVS_GITHUB_TOKEN") != null)
-            GITHUB_AUTH_TOKEN = System.getenv("LPVS_GITHUB_TOKEN");
     }
 
     private static LPVSQueue getGitHubWebhookConfig(GHRepository repo, GHPullRequest pR) {
@@ -148,13 +109,7 @@ public class LPVSDetectService {
                             Arrays.asList(pullRequestSplit)
                                     .subList(0, pullRequestSplit.length - 1));
             int pullRequestNum = Integer.parseInt(pullRequestSplit[pullRequestSplit.length - 1]);
-            if (GITHUB_AUTH_TOKEN == null || GITHUB_AUTH_TOKEN.isEmpty()) setGithubTokenFromEnv();
-            if (GITHUB_API_URL == null || GITHUB_API_URL.isEmpty())
-                gitHub = GitHub.connect(GITHUB_LOGIN, GITHUB_AUTH_TOKEN);
-            else
-                gitHub =
-                        GitHub.connectToEnterpriseWithOAuth(
-                                GITHUB_API_URL, GITHUB_LOGIN, GITHUB_AUTH_TOKEN);
+            GitHub gitHub = gitHubConnectionService.connectToGitHubApi();
             GHRepository repo = gitHub.getRepository(pullRequestRepo);
             GHPullRequest pR = repo.getPullRequest(pullRequestNum);
             return LPVSDetectService.getGitHubWebhookConfig(repo, pR);
