+++ src/main/java/com/lpvs/service/scanner/scanoss/LPVSScanossDetectService.java
@@ -35,8 +35,6 @@ public class LPVSScanossDetectService {
 
     @Autowired private LPVSLicenseService licenseService;
 
-    @Autowired private LPVSGitHubService gitHubService;
-
     @Autowired private LPVSLicenseRepository lpvsLicenseRepository;
 
     private Boolean debug;
@@ -45,11 +43,9 @@ public class LPVSScanossDetectService {
     public LPVSScanossDetectService(
             @Value("${debug:false}") Boolean debug,
             LPVSLicenseService licenseService,
-            LPVSGitHubService gitHubService,
             LPVSLicenseRepository lpvsLicenseRepository) {
         this.debug = debug;
         this.licenseService = licenseService;
-        this.gitHubService = gitHubService;
         this.lpvsLicenseRepository = lpvsLicenseRepository;
     }
 
@@ -83,7 +79,7 @@ public class LPVSScanossDetectService {
             int status = process.waitFor();
 
             if (status == 1) {
-                log.error("Scanoss scanner terminated with none-zero code. Terminating.");
+                log.error("Scanoss scanner terminated with non-zero code. Terminating.");
                 BufferedReader output = null;
                 try {
                     output =
@@ -91,13 +87,13 @@ public class LPVSScanossDetectService {
                                     new InputStreamReader(process.getErrorStream(), "UTF-8"));
                     log.error(output.readLine());
                     throw new Exception(
-                            "Scanoss scanner terminated with none-zero code. Terminating.");
+                            "Scanoss scanner terminated with non-zero code. Terminating.");
                 } finally {
                     if (output != null) output.close();
                 }
             }
         } catch (IOException | InterruptedException ex) {
-            log.error("Scanoss scanner terminated with none-zero code. Terminating.");
+            log.error("Scanoss scanner terminated with non-zero code. Terminating.");
             throw ex;
         }
